Fix typo, move tests to MapAttributions.test.tsx
